serv: Fixed extremely unlikely race condition in ReceiverChannel.
ReceiverChannel::connect() could establish a full connection and set
`receiverAvailable` to true and notify before ReceiverChannel::run()
even got a chance to sleep on the condition variable. This means that
run() could be sleeping while a connection just got established.

The fix is to initialize `localReceiverAvailable` to the actual value
of `receiverAvailable` instead of assuming the latter is false at this
point.

The only way this could occur is if, immediately after the Server
started, the Receiver established a full connection before the
ReceiverChannel thread ever got scheduled to run. The consequence of
this is that the Receiver's first connection would time out, but future
connections would be fine.